+++ src/main/java/com/example/hotdeal/domain/common/client/product/ProductApiClient.java
@@ -1,6 +1,6 @@
 package com.example.hotdeal.domain.common.client.product;
 
-import com.example.hotdeal.domain.product.product.domain.dto.SearchProductResponse;
+import com.example.hotdeal.domain.common.client.product.dto.SearchProductResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
