+++ src/main/java/com/example/hotdeal/domain/notification/api/NotificationController.java
@@ -1,0 +1,4 @@
+package com.example.hotdeal.domain.notification.api;
+
+public class NotificationController {
+}
