+++ src/main/java/com/example/hotdeal/domain/order/application/Service/OrderService.java
@@ -7,9 +7,9 @@ import com.example.hotdeal.domain.order.domain.Order;
 import com.example.hotdeal.domain.order.domain.OrderItem;
 import com.example.hotdeal.domain.order.enums.OrderStatus;
 import com.example.hotdeal.domain.order.infra.OrderRepository;
-import com.example.hotdeal.domain.product.product.domain.Product;
-import com.example.hotdeal.domain.product.product.domain.dto.SearchProductResponse;
-import com.example.hotdeal.domain.product.product.infra.ProductRepositoryImpl;
+import com.example.hotdeal.domain.product.domain.Product;
+import com.example.hotdeal.domain.common.client.product.dto.SearchProductResponse;
+import com.example.hotdeal.domain.product.infra.ProductRepositoryImpl;
 import com.example.hotdeal.global.enums.CustomErrorCode;
 import com.example.hotdeal.global.exception.CustomException;
 import lombok.AllArgsConstructor;
