+++ src/main/java/com/example/hotdeal/domain/product/application/ProductService.java
@@ -1,0 +1,15 @@
+package com.example.hotdeal.domain.product.application;
+
+import com.example.hotdeal.domain.product.domain.Product;
+import com.example.hotdeal.domain.product.domain.dto.CreateProductRequest;
+import com.example.hotdeal.domain.product.domain.dto.CreateProductResponse;
+import com.example.hotdeal.domain.common.client.product.dto.SearchProductResponse;
+
+import java.util.List;
+
+public interface ProductService {
+    CreateProductResponse createProduct(CreateProductRequest request);
+    Product findById(Long productId);
+    void deleteProduct(Long productId);
+    List<SearchProductResponse> searchProductsByIds(List<Long> productIds);
+}
