+++ src/main/java/com/example/hotdeal/domain/product/exception/ProductExceptionHandler.java
@@ -1,0 +1,39 @@
+package com.example.hotdeal.domain.product.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+
+@RestControllerAdvice(basePackages = "com.example.hotdeal.domain.product")
+public class ProductExceptionHandler {
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<Map<String, Object>> handleValidationException(MethodArgumentNotValidException e) {
+        BindingResult bindingResult = e.getBindingResult();
+        String message = Objects.requireNonNull(bindingResult.getFieldError()).getDefaultMessage();
+
+        Map<String, Object> response = new HashMap<>();
+        response.put("error", "VALIDATION_ERROR");
+        response.put("message", message);
+        response.put("status", HttpStatus.BAD_REQUEST.value());
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
+    }
+
+    @ExceptionHandler(ProductException.class)
+    public ResponseEntity<Map<String, Object>> handleProductException(ProductException e) {
+        Map<String, Object> response = new HashMap<>();
+        response.put("error", e.getErrorCode());
+        response.put("message", e.getMessage());
+        response.put("status", e.getHttpStatus().value());
+
+        return ResponseEntity.status(e.getHttpStatus()).body(response);
+    }
+}
