+++ src/main/java/com/example/hotdeal/domain/product/infra/ProductRepository.java
@@ -1,0 +1,14 @@
+package com.example.hotdeal.domain.product.infra;
+
+import com.example.hotdeal.domain.product.domain.Product;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface ProductRepository {
+    Product save(Product product);
+    List<Product> saveAll(List<Product> products);
+    Optional<Product> findById(Long productId);
+    void delete(Product product);
+    List<Product> findAllByIds(List<Long> productIds);
+}
