--- src/main/java/com/example/hotdeal/domain/product/notification/api/NotificationController.java
@@ -1,4 +1,0 @@
-package com.example.hotdeal.domain.product.notification.api;
-
-public class NotificationController {
-}
