--- src/main/java/com/example/hotdeal/domain/product/notification/infra/NotificationRepository.java
@@ -1,7 +1,0 @@
-package com.example.hotdeal.domain.product.notification.infra;
-
-import com.example.hotdeal.domain.product.notification.domain.Notification;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {
-}
