--- src/main/java/com/example/hotdeal/domain/product/product/domain/dto/UpdateProductResponse.java
@@ -1,22 +1,0 @@
-package com.example.hotdeal.domain.product.product.domain.dto;
-
-import com.example.hotdeal.domain.product.product.domain.Product;
-import lombok.Getter;
-
-import java.math.BigDecimal;
-
-@Getter
-public class UpdateProductResponse {
-
-    private final Long productId;
-    private final String productName;
-    private final String productDescription;
-    private final BigDecimal productPrice;
-
-    public UpdateProductResponse(Product product) {
-        this.productId = product.getProductId();
-        this.productName = product.getProductName();
-        this.productDescription = product.getProductDescription();
-        this.productPrice = product.getProductPrice();
-    }
-}
