+++ src/main/java/com/example/hotdeal/domain/stock/infra/StockRepository.java
@@ -18,5 +18,7 @@ public interface StockRepository extends JpaRepository<Stock, Long> {
     @Lock(LockModeType.PESSIMISTIC_WRITE)
     @Query("select s from Stock s where s.id = :id")
     Optional<Stock> findByIdWithPessimisticLock(@Param("id") Long id);
+
+    Optional<Stock> findByProductId(Long productId);
 }
