+++ src/test/java/com/example/hotdeal/ProductDataInsertTest.java
@@ -1,8 +1,8 @@
 package com.example.hotdeal;
 
-import com.example.hotdeal.domain.product.product.domain.Product;
-import com.example.hotdeal.domain.product.product.domain.ProductCategory;
-import com.example.hotdeal.domain.product.product.infra.ProductRepository;
+import com.example.hotdeal.domain.product.domain.Product;
+import com.example.hotdeal.domain.product.domain.ProductCategory;
+import com.example.hotdeal.domain.product.infra.ProductRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
